requirements.txt
Django==2.2.2
django-environ==0.4.5
django-filter==2.1.0
django-model-utils==3.1.2
djangorestframework==3.9.4
Markdown==3.1.1
Pillow==6.0.0
psycopg2==2.8.2
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
